Fix compiler warning that crept in with draw=from_caller
ParseDrawCommand checks to ensure that the entire value from the ini
file has been consumed in common code, and bails otherwise. For the
draw=from_caller case I set the read length to the end of the string
since wcscmp has already given this guarantee, but the return type of
string::length is different to %n used in the scanf functions. Cast it
to remove the warning.

+++ DirectX11/CommandList.cpp
@@ -153,7 +153,7 @@ static bool ParseDrawCommand(const wchar_t *key, wstring *val,
 	if (!wcscmp(key, L"draw")) {
 		if (!wcscmp(val->c_str(), L"from_caller")) {
 			operation->type = DrawCommandType::FROM_CALLER;
-			end = val->length();
+			end = (int)val->length();
 		} else {
 			operation->type = DrawCommandType::DRAW;
 			nargs = swscanf_s(val->c_str(), L"%u, %u%n", &operation->args[0], &operation->args[1], &end);
